client added

+++ Desktop/BTPTunisia-master/server/src/controllers/produit.controller.js
@@ -2,7 +2,20 @@ const Produit = require('../models/produit.model')
 
 const produitController = {};
 
-// Get products
+// Get all products
+produitController.getAll = async (req, res, next) => {
+    try {
+        const produits = await Produit.find();
+        return res.send({
+            produits
+        })
+
+    } catch (err) {
+        next(err);
+    }
+}
+
+// Get owner products
 produitController.get = async (req, res, next) => {
     const { moe } = req;
 
@@ -62,6 +75,13 @@ produitController.delete = async (req, res, next) => {
     const { _id } = req.params;
 
     try {
+        const check= await Produit.findOne({_id})
+        if(!check.owner.equals(req.moe._id)){
+            const err= new Error('Suppression refusée, ce produit ne vous appartient pas!')
+            err.status=401
+            throw err            
+        }
+
         await Produit.deleteOne({ _id })
         return res.send({
             message: "Produit supprimé"
@@ -76,12 +96,20 @@ produitController.update = async (req, res, next) => {
     const { _id } = req.params;
    
     try {
-        await Produit.findOneAndUpdate(
+        const check= await Produit.findOne({_id})
+        if(!check.owner.equals(req.moe._id)){
+            const err= new Error('Modification refusée, ce produit ne vous appartient pas!')
+            err.status=401
+            throw err            
+        }
+
+        const produitUpdated = await Produit.findOneAndUpdate(
             { _id }, 
             { $set: req.body }
         )
         return res.send({
-            message: "Produit modifié"
+            message: "Produit modifié",
+            produitUpdated
         })
     } catch (err) {
         next(err);
